应该是float。

+++ Chassis/Inc/Chassis_task.h
@@ -10,7 +10,7 @@ typedef struct
     uint16_t can_id;        // ID号
     int16_t set_current;    // 发送信息
     uint16_t rotor_angle;   // 现在的角度
-    uint16_t real_angle;    // 实际角度
+    fp32 real_angle;        // 实际角度
     int16_t rotor_speed;    // 现在的转速
     int16_t torque_current; // 实际转矩电流
     uint8_t temp;           // 电机温度
